requirements.txt
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
more-itertools==8.0.2
mysql-connector==2.2.9
mysql-connector-python==8.0.18
numpy==1.17.4
oauthlib==3.1.0
prettyprint==0.1.5
protobuf==3.11.0
PyMySQL==0.9.3
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
six==1.13.0
SQLAlchemy==1.3.11
tweepy==3.8.0
urllib3==1.25.7
Werkzeug==0.16.0
WTForms==2.2.1
